Reorganised the copdebase
Reorganised the file structure architecture of the code base.
Different classes. modules etc. were moved where they belong.
This way I achieved a more clean code base, easier to understand and navigate.
