+++ Server/WEEE-Calculator-Server/Infrastructure/Loggers/Logger.ts
