+++ Server/WEEE-Calculator-Server/Infrastructure/Utilities/RequestUtils .ts
