+++ Server/WEEE-Calculator-Server/Modules/Collection/Models/Collection.ts
