+++ Server/WEEE-Calculator-Server/Modules/Product/Models/Product.ts
