+++ Server/WEEE-Calculator-Server/ServiceLayer/Services/CollectionsCalculator.ts
@@ -1,9 +1,8 @@
 import CollectionsTotalWeightMap from "../../BLOC/CollectionsTotalWeighMap";
-import CollectionsDAO from "../../DAOs/CollectionsDAO";
-import CollectionsGraphDAO from "../../DAOs/CollectionsGraphDAO";
-import OrdersDAO from "../../DAOs/OrdersDAO";
-import CollectionsManager from "./CollectionsManager";
-import OrdersManager from "./OrdersManager";
+import CollectionsDAO from "../../Modules/Collection/DAOs/CollectionsDAO";
+import CollectionsGraphDAO from "../../Modules/Collection/DAOs/CollectionsGraphDAO";
+import CollectionsManager from "../../Modules/Collection/Services/CollectionsManager";
+import OrdersManager from "../../Modules/Order/Services/OrdersManager";
 
 /**
  * A service class with a signle purpose to calculate the collections total weight and return it
