+++ Server/WEEE-Calculator-Server/main.ts
@@ -5,37 +5,33 @@ import dotenv from "dotenv";
 import webHookRouter from "./Routes/webhooksRouter";
 
 //Services Imports
-import CollectionsTotalWeightMap from "./BLOC/CollectionsTotalWeighMap";
-import CollectionsManager from "./ServiceLayer/Services/CollectionsManager";
-import ProductsManager from "./ServiceLayer/Services/ProductsManager";
 import CollectionsCalculator from "./ServiceLayer/Services/CollectionsCalculator";
 
 //DAO Factory
 import DaoFactory from "./Factory/DaoFactory";
 
 //DAO imports
-import ProductsDAO from "./DAOs/ProductsDAO";
-import CollectionsGraphDAO from "./DAOs/CollectionsGraphDAO";
-import CollectionsDAO from "./DAOs/CollectionsDAO";
-import OrdersDAO from "./DAOs/OrdersDAO";
+import CollectionsGraphDAO from "./Modules/Collection/DAOs/CollectionsGraphDAO";
+import CollectionsDAO from "./Modules/Collection/DAOs/CollectionsDAO";
+import OrdersDAO from "./Modules/Order/DAOs/OrdersDAO";
 
 //Utilities Imports
-import RequestUtils from "./Utilities/RequestUtils ";
-import CollectionProductService from "./ServiceLayer/Services/CollectionsProductService";
-import ResourceNotFound from "./ExceptionModels/ResourceNotFoundException";
-import Collection from "./Models/Collection";
-import Product from "./Models/Product";
-import OrdersManager from "./ServiceLayer/Services/OrdersManager";
-import OrdersGraphDAO from "./DAOs/OrdersGraphDAO";
+import RequestUtils from "./Infrastructure/Utilities/RequestUtils ";
+import OrdersManager from "./Modules/Order/Services/OrdersManager";
+import OrdersGraphDAO from "./Modules/Order/DAOs/OrdersGraphDAO";
 import {
   routeErrorLogger,
   errorLogger,
   routeResponseLogger,
-} from "./Helpers/Logger";
+} from "./Infrastructure/Loggers/Logger";
 import healthRouter from "./Routes/HealthRouter";
 
+//Router imports
+import collectionRouter from "./Modules/Collection/Routers/collectionRouter";
+import productRouter from "./Modules/Product/Routers/productRouter";
+
 //Queue Workers
-import { shopRedactWorker } from "./Workers/webhooks_queue_worker";
+import { shopRedactWorker } from "./Infrastructure/Workers/webhooks_queue_worker"; //Don't erase this line
 
 dotenv.config();
 
@@ -132,6 +128,11 @@ app.use("/health", healthRouter);
 
 //========= Web hook routes ======== //
 app.use("/gdpr-compliance/webhooks", webHookRouter);
+
+//========= Web API routes ======== //
+app.use("/api/v1", collectionRouter);
+
+app.use("/api/v1", productRouter);
 
 app.post("/api/v1/initCalculation", async (req: Request, res: Response) => {
   const route: string = "/initCalculation";
@@ -238,523 +239,4 @@ app.post("/api/v1/initCalculation", async (req: Request, res: Response) => {
     res.status(500).send(`Internal server error`);
     return;
   }
-});
-
-/**
- * This route is designated for creating collections
- * The route expects headers with string accessToken and string hostName
- * The rout expects to get an array of Maps containing collection title as key and collection description as value:
- * [
-      {
-        'collectionTitle': "collectionDescription"
-      },
-      {
-        'collectionTitle': "collectionDescription"
-      }, 
-    ]
- */
-app.post("/api/v1/createCollection", async (req: Request, res: Response) => {
-  const route: string = "/createCollection";
-  try {
-    console.log("============= \n/createCollection requested by: ip ", req.ip);
-    const { accessToken, hostName } = RequestUtils.extractHeaders(req);
-
-    if (!accessToken || !hostName) {
-      routeErrorLogger(route, req, "Missing headers", 400);
-
-      res.status(400).send("Missing headers");
-      return;
-    }
-
-    const collections = req.body;
-    console.log("collections received: ", collections);
-
-    const collectionsMapsArray: Array<Map<string, string>> = collections.map(
-      (obj) => new Map(Object.entries(obj))
-    );
-
-    //Initialising the DAO factory class
-    const daoFactory = new DaoFactory(accessToken, hostName);
-    const collectionsRestDao: CollectionsDAO =
-      daoFactory.getDAO("collectionsRestDao");
-    const collectionsGraphDao: CollectionsGraphDAO = daoFactory.getDAO(
-      "collectionsGraphDao"
-    );
-
-    //initialising the collectionsManager class
-    const collectionsManager: CollectionsManager = new CollectionsManager(
-      collectionsGraphDao,
-      collectionsRestDao
-    );
-    const result: { isSuccess: boolean; error?: string } =
-      await collectionsManager.createCollectionsFor(collectionsMapsArray);
-
-    if (result.error) {
-      routeErrorLogger(route, req, result.error, 500);
-
-      res.status(500).send("Error creating collections. Internal server error");
-      return;
-    }
-
-    if (result.isSuccess) {
-      routeResponseLogger(route, req, "Collections created successfully", 201);
-
-      res.status(201).send("Collections created");
-      return;
-    } else {
-      routeErrorLogger(
-        route,
-        req,
-        result.error ? result.error : "Action unsuccessful",
-        500
-      );
-
-      res.status(500).send("Error creating collections");
-      return;
-    }
-  } catch (e) {
-    routeErrorLogger(route, req, "Internal server error", 500);
-
-    res.status(500).send(`Internal server error`);
-    return;
-  }
-});
-
-/**
- * This route is designated for getting all WEEE collections from vendor's store
- * The route expects headers with string accessToken and string hostName
- */
-app.get("/api/v1/weeeCollections/all", async (req, res) => {
-  const route: string = "/weeeCollections/all";
-  try {
-    console.log(
-      "============= \n/weeeCollecations/all requested by: IP ",
-      req.ip
-    );
-
-    const { accessToken, hostName } = RequestUtils.extractHeaders(req);
-
-    if (!accessToken || !hostName) {
-      routeErrorLogger(route, req, "Missing headers", 400);
-
-      res.status(400).send("Missing headers");
-      return;
-    }
-    //Initialising the DAO factory class
-    const daoFactory: DaoFactory = new DaoFactory(accessToken, hostName);
-
-    //Getting the needed DAOs
-    const collectionsRestDao: CollectionsDAO =
-      daoFactory.getDAO("collectionsRestDao");
-    const collectionsGraphDao: CollectionsGraphDAO = daoFactory.getDAO(
-      "collectionsGraphDao"
-    );
-
-    //initialising the collectionsManager class
-    const collectionsManager: CollectionsManager = new CollectionsManager(
-      collectionsGraphDao,
-      collectionsRestDao
-    );
-
-    const result: {
-      isSuccess: boolean;
-      collections: Array<Collection>;
-      error?: string;
-    } = await collectionsManager.getWeeeCollections();
-
-    if (result.error) {
-      routeErrorLogger(route, req, result.error, 500);
-
-      res
-        .status(500)
-        .send(`Error getting weee collections. Internal server error`);
-      return;
-    }
-
-    if (result.isSuccess) {
-      if (result.collections && result.collections.length <= 0) {
-        routeResponseLogger(
-          route,
-          req,
-          "No WEEE collections found in vendor's store",
-          404
-        );
-
-        res.status(404).send("No WEEE collections found");
-        return;
-      }
-      routeResponseLogger(
-        route,
-        req,
-        "WEEE collections retreived successfully",
-        200
-      );
-
-      res.status(200).send(JSON.stringify(result.collections));
-      return;
-    }
-
-    routeErrorLogger(route, req, "Internal server error", 500);
-
-    res
-      .status(500)
-      .send(`Error getting all weee collections. Internal server error`);
-    return;
-  } catch (e) {
-    routeErrorLogger(route, req, e, 500);
-
-    res
-      .status(500)
-      .send(`Error getting all weee collections. Internal server error`);
-    return;
-  }
-});
-
-/**
- * This route is designated for getting all products belonging to a  collection
- * The route expects headers with string accessToken and string hostName
- * The route expects to get a collection id url parameter. The parameter is the id of the collection which products need to be fetched.
- */
-app.get("/api/v1/collection/:id/products/all", async (req, res) => {
-  const route: string = "/collection/:id/products/all";
-  try {
-    console.log(
-      "============= \n/collection/:id/products/all requested by: IP ",
-      req.ip
-    );
-
-    const { accessToken, hostName } = RequestUtils.extractHeaders(req);
-    const collectionId = Number(req.params.id);
-
-    if (!accessToken || !hostName) {
-      routeErrorLogger(route, req, "Missing headers", 400);
-
-      res.status(400).send("Missing headers");
-      return;
-    }
-
-    //Initialising the DAO factory class
-    const daoFactory: DaoFactory = new DaoFactory(accessToken, hostName);
-
-    //Getting the needed DAOs
-    const productsDAO: ProductsDAO = daoFactory.getDAO("productsDao");
-
-    //initialising the collectionsManager class
-    const productsManager: ProductsManager = new ProductsManager(productsDAO);
-
-    const result: { isSuccess: boolean; products: Product[]; error?: string } =
-      await productsManager.getProductsForCollection(collectionId);
-
-    if (result.error) {
-      routeErrorLogger(route, req, result.error, 500);
-
-      //Returns 500 if error occured
-      res
-        .status(500)
-        .send(
-          "Error getting collections's all products. Internal server error"
-        );
-    }
-
-    if (result.isSuccess) {
-      if (result.products.length == 0) {
-        //Still successful operation, but no products found
-        routeResponseLogger(
-          route,
-          req,
-          `No products were found that belong to collection ${collectionId} in vendors store.`,
-          404
-        );
-
-        res
-          .status(404)
-          .send(
-            `No products were found that belong to collection ${collectionId} in vendors store.`
-          );
-        return;
-      }
-      routeResponseLogger(
-        route,
-        req,
-        "Collection products retreived successfully",
-        200
-      );
-
-      res.status(200).send(result.products);
-      return;
-    }
-
-    routeErrorLogger(route, req, "Internal server error", 500);
-
-    //Returns 500 if error occured
-    res
-      .status(500)
-      .send("Error getting collections's all products. Internal server error");
-  } catch (e) {
-    routeErrorLogger(route, req, e, 500);
-
-    res
-      .status(500)
-      .send(`Error getting collections's all products. Internal server error`);
-    return;
-  }
-});
-
-/**
- * This route is designated for getting all products
- * The route expects headers with string accessToken and string hostName
- * The route sends back an array of product objects
- */
-app.get("/api/v1/products/all", async (req: Request, res) => {
-  const route: string = "/api/v1/products/all";
-  try {
-    console.log("============= \n/products/all  requested by: IP ", req.ip);
-
-    const { accessToken, hostName } = RequestUtils.extractHeaders(req);
-
-    if (!accessToken || !hostName) {
-      routeErrorLogger(route, req, "Missing headers", 400);
-
-      res.status(400).send("Missing headers");
-      return;
-    }
-    const daoFactory: DaoFactory = new DaoFactory(accessToken, hostName);
-    const productsDao: ProductsDAO = daoFactory.getDAO("productsDao");
-
-    const productManager = new ProductsManager(productsDao);
-    const result = await productManager.getAllActiveProducts();
-
-    if (result.error) {
-      routeErrorLogger(route, req, result.error, 500);
-
-      res.status(500).send("Error getting all products. Internal server error");
-      return;
-    }
-
-    if (result.isSuccess) {
-      if (result.products.length == 0) {
-        routeResponseLogger(
-          route,
-          req,
-          `No products were found in vendors store.`,
-          404
-        );
-
-        res.status(404).send("No products were found in vendors store.");
-        return;
-      }
-      routeResponseLogger(
-        route,
-        req,
-        "All products retreived successfully",
-        200
-      );
-
-      res.status(200).send(result.products);
-      return;
-    } else {
-      routeErrorLogger(route, req, "Internal server error", 500);
-
-      res.status(500).send("Error getting all products. Internal server error");
-      return;
-    }
-  } catch (e) {
-    routeErrorLogger(route, req, e, 500);
-
-    res.status(500).send(`Error getting all products. Internal server error`);
-    return;
-  }
-});
-
-/**
- * This route is designated for adding products to a collection
- * The route expects headers with string accessToken and string hostName, and a JSON body with:
- * {
- *   "collection": "collectionId",
- *   "products": ["productId", "productId", ...]
- * }
- * collection is the collection id of the collection that products will be added to
- * products is an array of product ids that will be added to the collection
- */
-app.post(
-  "/api/v1/addProductsToCollection",
-  async (req: Request, res: Response) => {
-    const route: string = "/addProductsToCollection";
-    try {
-      console.log(
-        "============= \n/addProductsToCollection  requested by: IP ",
-        req.ip
-      );
-
-      const { accessToken, hostName } = RequestUtils.extractHeaders(req);
-
-      if (!accessToken || !hostName) {
-        routeErrorLogger(route, req, "Missing headers", 400);
-
-        res.status(400).send({ message: "Missing headers" });
-        return;
-      }
-
-      const collectionId: string = req.body.collection;
-      const products: Array<string> = req.body.products;
-
-      if (typeof collectionId !== "string" || !Array.isArray(products)) {
-        routeErrorLogger(route, req, "Parameters of wrong type", 400);
-
-        res.status(400).send({ message: "Prameters are not of correct type" });
-        return;
-      }
-
-      const daoFactory: DaoFactory = new DaoFactory(accessToken, hostName);
-      const collectionsRestDao: CollectionsDAO =
-        daoFactory.getDAO("collectionsRestDao");
-      const collectionsGraphDao: CollectionsGraphDAO = daoFactory.getDAO(
-        "collectionsGraphDao"
-      );
-      const collectionsProductService: CollectionProductService =
-        new CollectionProductService(collectionsGraphDao, collectionsRestDao);
-
-      const result = await collectionsProductService.addProductsToCollection(
-        collectionId,
-        products
-      );
-
-      if (result.error) {
-        routeErrorLogger(route, req, result.error, 500);
-
-        res
-          .status(500)
-          .send("Error adding products to collection. Internal server error");
-        return;
-      }
-
-      if (result.isSuccess) {
-        routeResponseLogger(route, req, "Products added successfully", 200);
-
-        res
-          .status(200)
-          .send({ message: "Products successfully added to collecton" });
-        return;
-      } else {
-        routeErrorLogger(route, req, "Internal server error", 500);
-
-        res
-          .status(500)
-          .send({ message: "Error adding products to collection" });
-        return;
-      }
-    } catch (err) {
-      if (err instanceof ResourceNotFound) {
-        routeErrorLogger(route, req, err, 400);
-
-        res.status(400).send(err);
-        return;
-      } else {
-        routeErrorLogger(route, req, err, 500);
-
-        res
-          .status(500)
-          .send(`Error adding products to collection. Internal server error`);
-        return;
-      }
-    }
-  }
-);
-
-/**
- * This route is designated for adding products to a collection
- * The route expects headers with string accessToken and string hostName, and a JSON body with:
- * {
- *   "collection": "collectionId",
- *   "products": ["productId", "productId", ...]
- * }
- * collection is the collection id of the collection that products will be removed from
- * products is an array of product ids that will be removed from the collection
- */
-app.post(
-  "/api/v1/removeProductsFromCollection",
-  async (req: Request, res: Response) => {
-    const route: string = "/removeProductsFromCollection";
-    try {
-      console.log(
-        "============= \n/removeProductsFromCollection  requested by: IP ",
-        req.ip
-      );
-
-      const { accessToken, hostName } = RequestUtils.extractHeaders(req);
-
-      if (!accessToken || !hostName) {
-        routeErrorLogger(route, req, "Missing headers", 500);
-
-        res.status(400).send("Missing headers");
-        return;
-      }
-
-      const collectionId = req.body.collection;
-      const products = req.body.products;
-
-      if (typeof collectionId !== "string" || !Array.isArray(products)) {
-        routeErrorLogger(route, req, "Parameters of wrong type", 500);
-
-        res.status(400).send("Prameters are not of correct type");
-        return;
-      }
-
-      const daoFactory: DaoFactory = new DaoFactory(accessToken, hostName);
-      const collectionsRestDao: CollectionsDAO =
-        daoFactory.getDAO("collectionsRestDao");
-      const collectionsGraphDao: CollectionsGraphDAO = daoFactory.getDAO(
-        "collectionsGraphDao"
-      );
-      const collectionsProductService: CollectionProductService =
-        new CollectionProductService(collectionsGraphDao, collectionsRestDao);
-
-      const result: { isSuccess: boolean; error?: string } =
-        await collectionsProductService.removeProductsFromCollection(
-          collectionId,
-          products
-        );
-
-      if (result.error) {
-        routeErrorLogger(route, req, result.error, 500);
-
-        res.status(500).send("Error removing products. Internal server error");
-        return;
-      }
-      if (result.isSuccess) {
-        routeResponseLogger(route, req, "Products removed successfully", 200);
-
-        res
-          .status(200)
-          .send({ message: "Products successfully removed from collecton" });
-        return;
-      } else {
-        routeErrorLogger(route, req, "Internal server error", 500);
-
-        res
-          .status(500)
-          .send({ message: "Error removing products from collection" });
-        return;
-      }
-    } catch (err) {
-      if (err instanceof ResourceNotFound) {
-        res.status(400).send(err);
-        return;
-      } else {
-        routeErrorLogger(route, req, err, 500);
-
-        res
-          .status(500)
-          .send(
-            `Error removing products from collection. Internal server error`
-          );
-        return;
-      }
-    }
-  }
-);
-
-app.get("/api/v1/health", async (req: Request, res: Response) => {
-  console.log("============= \nHealth check was requested by ip: ", req.ip);
-  res.status(200).send("App is healthy");
 });
